Allow the user to pass an arbitrary useragent object.
The purpose behind this is to make it easy to mock network calls in a
test environment.  This is very easy to do with something like
Test::LWP::UserAgent, but more difficult to do with Furl.  The
difference between the two classes is one argument to the post() method.
Furl needs the headers passes as an undef and with LWP::UserAgent, you
can just omit the headers entirely.